requirements.txt
# Core dependencies
crewai==0.28.0
langchain>=0.1.10,<0.2.0
langchain-openai==0.0.5
python-dotenv==1.0.0
colorama==0.4.6
rich==13.7.0 